Module type Trie.Intf

type path

type of path point

type 'a node

type of trie node

val create : 'a option -> 'a node

create a new trie tree with an optional element

val get : 'a node -> path -> 'a option

returns the value associated with the path

val set : 'a node -> path -> 'a -> unit

associate the value with the path

val unset : 'a node -> path -> unit

remove an association of the path

val sub : 'a node -> path -> 'a node option

returns the sub node associated with the path

val is_leaf : 'a node -> bool

returns whether the node is a leaf of the tree